+++ lib/src/models/auth_cache_data.dart
@@ -1,19 +1,21 @@
 import 'dart:convert';
 
+import 'package:equatable/equatable.dart';
+
 import 'package:json_annotation/json_annotation.dart';
 
 part 'auth_cache_data.g.dart';
 
 ///鉴权信息数据对象
 @JsonSerializable()
-class AuthCacheData {
+class AuthCacheData extends Equatable {
   final String userID;
 
   final String imToken;
 
   final String? chatToken;
 
-  AuthCacheData({required this.userID, required this.imToken, this.chatToken});
+  const AuthCacheData({required this.userID, required this.imToken, this.chatToken});
 
   factory AuthCacheData.fromJson(Map<String, dynamic> json) => _$AuthCacheDataFromJson(json);
 
@@ -23,4 +25,7 @@ class AuthCacheData {
   String toString() {
     return jsonEncode(toJson());
   }
+
+  @override
+  List<Object?> get props => [userID, imToken, chatToken];
 }